+++ internal/cli/atlas/clusters/clusters.go
@@ -16,7 +16,7 @@ package clusters
 
 import (
 	"github.com/mongodb/mongocli/internal/cli"
-	"github.com/mongodb/mongocli/internal/cli/atlas/clusters/connetionstring"
+	"github.com/mongodb/mongocli/internal/cli/atlas/clusters/connectionstring"
 	"github.com/mongodb/mongocli/internal/cli/atlas/clusters/indexes"
 	"github.com/mongodb/mongocli/internal/cli/atlas/onlinearchive"
 	"github.com/mongodb/mongocli/internal/cli/atlas/search"
@@ -44,7 +44,7 @@ func Builder() *cobra.Command {
 		indexes.Builder(),
 		search.Builder(),
 		onlinearchive.Builder(),
-		connetionstring.Builder())
+		connectionstring.Builder())
 
 	return cmd
 }
