+++ internal/cli/atlas/clusters/connectionstring/describe.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package connetionstring
+package connectionstring
 
 import (
 	"github.com/mongodb/mongocli/internal/cli"
