+++ internal/cli/atlas/clusters/connectionstring/description.go
@@ -11,22 +11,9 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+package connectionstring
 
-package connetionstring
-
-import (
-	"github.com/mongodb/mongocli/internal/cli"
-	"github.com/spf13/cobra"
+const (
+	connectionString = "Manage MongoDB cluster connection string"
+	describe         = "Retrieve the connection string of your MongoDB cluster."
 )
-
-func Builder() *cobra.Command {
-	const use = "connectionString"
-	cmd := &cobra.Command{
-		Use:     use,
-		Aliases: append(cli.GenerateAliases(use), "cs"),
-		Short:   connectionString,
-	}
-	cmd.AddCommand(DescribeBuilder())
-
-	return cmd
-}
